Update (add webpack and other configurations, add components)
