+++ .eslintrc.yml
@@ -1,6 +1,10 @@
+parserOptions:
+    ecmaVersion: 2015
+    sourceType: 'module'
 env:
   es6: true
   browser: true
   commonjs: true
+  node: true
 
 extends: 'htmlacademy/es6'
