+++ src/components/show-more.js
@@ -1,0 +1,2 @@
+export const createShowMoretempalte = () =>
+`<button class="films-list__show-more">Show more</button>`;
